Per-device connection toggle; lots more
Awesome progress!
Devices can be connected/disconnected as the button for each is now a
toggle. Turns red when connected. "Connecting..." string when connection
is initiated but not complete. Only shows devices containing "Blume" or
"Bluno" in the name. Color controls show/hide if any devices are
connected.
Added reload button to the bottom; this can be used to reload the app
completely in case of bugs.
Lots of cleanup.
Count pendingSends to know when sending to all connected devices is done
(and thus one last comand can be sent).