Device-specific commands
sendCommand(), sendAsk(), and sendData() now take a first argument
"devices", which can be null (to send to all devices), a single device,
or an array of devices. This lets us ask for device dimensions more
granularly and sendImage() with proper dimensions for each device.
Some other debug stuff and tweaks.