Universal demo mode
After 3 seconds on any platform, show a fake Blume device that can be
connected, showing the rest of the UI. Now the app can always be demoed.
This means various tweaks.
Released it using CodePush!
And resign ourselves to Cordova's reordering/tweaking of config.xml
plugin definitions, for now at least. It'll probably decide on a
different order later.