Support 2-dimensional image mode
This extends POV functionality to 2-dimensional devices.
Currently only 1 2D frame is supported, but the app could be modified to
send multiple frames of a GIF.
I realized that Adafruit's DotStart FeatherWing's 0,0 is lower-left,
which is upside-down from my perspective, so I add FLIP_HORIZONTAL and
FLIP_VERTICAL to account for such cases.
This makes text rendering more straightforward (so *that's* why I needed
to reverse the y!) and lets us standardize the pixel order that the app
sends for imagery.
For POV the app still sends pixel data "upside down" I think;
I'm punting on this for now to support older devices that expect it this
way.