Add 24-bit color support
8-bit color was necessary for the Bluno Beetle which has limited SRAM
and lower data rates, but 24-bit is fine with the ESP32. Now the app
decides which format to send based on MAX_FRAMES and the device supports
both bit depths.
While testing the old Bluno POV devices I noticed that MAX_FPS does
weird stuff for them, so just try to run them as fast as possible, which
is fine for the slow Arduino core.